+++ src/encoded/item_utils/library.py
@@ -1,4 +1,7 @@
-from typing import Any, Dict, Union
+from typing import Any, Dict, List, Optional, Union
+
+from . import analyte as analyte_utils
+from .utils import RequestHandler, get_property_value_from_identifier
 
 
 def get_assay(library: Dict[str, Any]) -> Union[str, Dict[str, Any]]:
@@ -9,3 +12,14 @@ def get_assay(library: Dict[str, Any]) -> Union[str, Dict[str, Any]]:
 def get_analyte(library: Dict[str, Any]) -> Union[str, Dict[str, Any]]:
     """Get analyte connected to library."""
     return library.get("analyte", "")
+
+
+def get_samples(
+    library: Dict[str, Any], request_handler: Optional[RequestHandler] = None
+) -> List[str]:
+    """Get samples connected to library."""
+    if request_handler:
+        return get_property_value_from_identifier(
+            request_handler, get_analyte(library), analyte_utils.get_samples
+        )
+    return []