+++ src/encoded/types/tissue.py
@@ -1,6 +1,14 @@
+from typing import List
+
 from snovault import collection, load_schema
 
 from .sample_source import SampleSource
+
+
+def _build_tissue_embedded_list() -> List[str]:
+    return [
+        "donor.external_id",
+    ]
 
 
 @collection(
@@ -14,4 +22,4 @@ from .sample_source import SampleSource
 class Tissue(SampleSource):
     item_type = "tissue"
     schema = load_schema("encoded:schemas/tissue.json")
-    embedded_list = []
+    embedded_list = _build_tissue_embedded_list()